+++ src/php/Lookup/Response/Person.php
@@ -1,12 +1,9 @@
 <?php
 /**
- * Wrapper for CANDDi Lookup
- * https://api.canddi.net
+ * This is our default response for our
+ * person_xxx_get lookups
  *
- * @TODO REFACTOR THIS TO a separate composer package
- * @TODO REPLACE THIS WITH PRESONLINKEDIN
- *
- * @author Tim Langley
+ * @author George Meadows
  **/
 
 namespace CanddiAi\Lookup\Response;
@@ -15,13 +12,16 @@ use CanddiAi\Traits\GetArrayValue as NS_traitArrayValue;
 
 class Person
 {
-    const KEY_BIO           = 'Bio';
-    const KEY_CONTACT       = 'ContactInfo';
-    const KEY_FORENAME      = 'FirstName';
-    const KEY_ROLE          = 'JobTitle';
-    const KEY_SURNAME       = 'LastName';
-    const KEY_PHOTO         = 'Photos';
-    const KEY_SOCIAL        = 'SocialProfiles';
+    const KEY_NAME = 'Name';
+    const KEY_FORENAME = 'FirstName';
+    const KEY_MIDDLE = 'MiddleName';
+    const KEY_SURNAME = 'LastName';
+    const KEY_EMAILS = 'EmailAddresses';
+    const KEY_PHONES = 'PhoneNumbers';
+    const KEY_ROLE = 'Employment';
+    const KEY_EDUCATION = 'Education';
+    const KEY_PHOTO = 'Photos';
+    const KEY_SOCIAL = 'SocialMedia';
 
     use NS_traitArrayValue;
 
@@ -37,7 +37,7 @@ class Person
         return $this->_getArrayValue(
             $this->_arrResponse,
             [
-                self::KEY_CONTACT,
+                self::KEY_NAME,
                 self::KEY_FORENAME
             ],
             null
@@ -48,31 +48,38 @@ class Person
         return $this->_getArrayValue(
             $this->_arrResponse,
             [
-                self::KEY_CONTACT,
+                self::KEY_NAME,
                 self::KEY_SURNAME
-            ],
-            null
-        );
-    }
-    public function getBio()
-    {
-        return $this->_getArrayValue(
-            $this->_arrResponse,
-            [
-                self::KEY_BIO
             ],
             null
         );
     }
     public function getRole()
     {
-        return $this->_getArrayValue(
+        $arrRoles =  $this->_getArrayValue(
             $this->_arrResponse,
             [
                 self::KEY_ROLE
             ],
-            null
+            []
         );
+
+        foreach ($arrRoles as $arrRole) {
+            if (isset($arrRole['IsPrimary']) && $arrRole['IsPrimary']) {
+                $itemRole = new Item\Role($arrRole);
+                break;
+            }
+        }
+
+        if (!isset($itemRole)) {
+            if (empty($arrRoles)) {
+                // Got no photos, explicitly return null
+                return null;
+            }
+            // Choose the first photo if we don't have a primary
+            $itemRole = new Item\Role($arrRoles[0]);
+        }
+        return $itemRole;
     }
     public function getPhotos()
     {
@@ -83,48 +90,8 @@ class Person
             ],
             []
         );
-        $arrReturn  = [];
 
-        //This is a horrible way of returning
-        // too many loops = slow code
-        // @TODO refactor with an Iterator
-        foreach ($arrPhotos as $arrPhoto) {
-            $arrReturn[] = new Item\Photo($arrPhoto);
-        }
-        return $arrReturn;
-    }
-    /**
-     * Finds the first photo with IsPrimary = true
-     *
-     * @return Item\Photo | null    The primary image (first photo is used if
-     *                                  primary is not present), or null if no
-     *                                  photos are present.
-     */
-    public function getPrimaryPhoto()
-    {
-        $arrPhotos  = $this->_getArrayValue(
-            $this->_arrResponse,
-            [
-                self::KEY_PHOTO
-            ],
-            []
-        );
-
-        foreach ($arrPhotos as $arrPhoto) {
-            if (isset($arrPhoto['IsPrimary']) && $arrPhoto['IsPrimary']) {
-                $itemPhoto = new Item\Photo($arrPhoto);
-                break;
-            }
-        }
-        if (!isset($itemPhoto)) {
-            if (empty($arrPhotos)) {
-                // Got no photos, explicitly return null
-                return null;
-            }
-            // Choose the first photo if we don't have a primary
-            $itemPhoto = new Item\Photo($arrPhotos[0]);
-        }
-        return $itemPhoto;
+        return $arrPhotos;
     }
     public function getSocialProfiles()
     {
@@ -144,5 +111,15 @@ class Person
             $arrReturn[] = new Item\Social($arrProfile);
         }
         return $arrReturn;
+    }
+    public function getEducation()
+    {
+        return $this->_getArrayValue(
+            $this->_arrResponse,
+            [
+                self::KEY_EDUCATION
+            ],
+            []
+        );
     }
 }
