+++ test/php/Lookup/Response/PersonTest.php
@@ -2,255 +2,137 @@
 
 namespace CanddiAi\Lookup\Response;
 
-class PersonTest
+class PersonTestLinkedIn
     extends \CanddiAi\TestCase
 {
-    private $response;
 
     private function _getTestData()
     {
-        return array(
-            "ContactInfo" => array(
-                "FirstName" => "Bart",
-                "LastName" => "Lorang",
-                "Websites" => array(
-                    array(
-                        "url" => "https:\/\/fullcontact.com\/bart"
-                    ),
-                    array(
-                        "url" => "http:\/\/bartlorang.com"
-                    )
-                ),
-                "Chats" => array(
-                    array(
-                        "client" => "gtalk",
-                        "handle" => "lorangb@gmail.com"
-                    ),
-                    array(
-                        "client" => "skype",
-                        "handle" => "bart.lorang"
-                    )
-                )
-            ),
-            "Demographics" => array(
-                "LocationGeneral" => "80202 Denver, Colorado, United States",
-                "LocationDeduced" => array(
-                    "NormalizedLocation" => "Denver, Colorado, United States",
-                    "DeducedLocation" => "Denver, Colorado, United States",
-                    "City" => array(
-                        "Deduced" => false,
-                        "Name" => "Denver"
-                    ),
-                    "State" => array(
-                        "Deduced" => false,
-                        "Name" => "Colorado",
-                        "Code" => "CO"
-                    ),
-                    "Country" => array(
-                        "Deduced" => false,
-                        "Name" => "United States",
-                        "Code" => "US"
-                    ),
-                    "Continent" => array(
-                        "Deduced" => true,
-                        "Name" => "North America"
-                    ),
-                    "County" => array(
-                        "Deduced" => true,
-                        "Name" => ""
-                    ),
-                    "Likelihood" => 1
-                ),
-                "Gender" => "Male"
-            ),
-            "SocialProfiles" => array(
-                array(
-                    "bio" => "Co-Founder and CEO of FullContact",
-                    "type" => "aboutme",
-                    "typeId" => "aboutme",
-                    "typeName" => "About.me",
-                    "url" => "https:\/\/about.me\/lorangb",
-                    "username" => "lorangb"
-                ),
-                array(
-                    "type" => "facebook",
-                    "typeId" => "facebook",
-                    "typeName" => "Facebook",
-                    "url" => "https:\/\/www.facebook.com\/bart.lorang"
-                ),
-                array(
-                    "type" => "github",
-                    "typeId" => "github",
-                    "typeName" => "Github",
-                    "url" => "https:\/\/github.com\/lorangb",
-                    "username" => "lorangb"
-                ),
-                array(
-                    "followers" => 1,
-                    "type" => "google",
-                    "typeId" => "google",
-                    "typeName" => "GooglePlus",
-                    "url" => "https:\/\/plus.google.com\/111748526539078793602",
-                    "id" => "111748526539078793602"
-                ),
-                array(
-                    "type" => "instagram",
-                    "typeId" => "instagram",
-                    "typeName" => "Instagram",
-                    "url" => "https:\/\/instagram.com\/bartlorang"
-                ),
-                array(
-                    "bio" => "Mr. Lorang is a proven entrepreneur, executive and manager in the global technology industry. Mr. Lorang is active in the startup technology community as an angel investor, strategic advisor and speaker at industry events. Bart serves as Co-Founder & CEO of FullContact. Mr. Lorang is responsible for communicating FullContact's vision and strategy. Mr. Lorang is a visionary technologist with extensive experience conceiving, designing, building, marketing and selling enterprise software solutions on a global scale. Bart is also Co-Founder and Managing Director of v1.vc, a $5M seed stage fund dedicated to helping crazy entrepreneurs change the world. Bart serves on the Board of the Colorado Technology Association, Rapt Media and is on the Advisory Board of Education Funding Partners. Bart is a regular guest on FOX Business channel and has been featured by ABC, CNN, FOX News, MSNBC, Forbes, FastCompany, Yahoo, Inc Magazine and TechCrunch. Prior to founding FullContact, Mr. Lorang was an owner in Dimension Technology Solutions where he served as President and oversaw all day to day operations, customer engagements, partner relations, product development, sales and marketing functions. Mr. Lorang is recognized for providing solutions that are simple and work reliably. He strongly believes in using technology to solve problems as opposed to using problems to demonstrate technology. Mr. Lorang holds a Bachelor of Science degree in Computer Science from the University of Colorado and an MBA from the Daniels College of Business at University of Denver. Specialties: Investments, Startups, Financial Analysis, Sales, Technical Sales, Implementations, System Integration, Project Management, Leadership",
-                    "followers" => 500,
-                    "following" => 500,
-                    "type" => "linkedin",
-                    "typeId" => "linkedin",
-                    "typeName" => "LinkedIn",
-                    "url" => "https:\/\/www.linkedin.com\/in\/bartlorang",
-                    "username" => "bartlorang",
-                    "id" => "8995706"
-                ),
-                array(
-                    "type" => "quora",
-                    "typeId" => "quora",
-                    "typeName" => "Quora",
-                    "url" => "http:\/\/www.quora.com\/bart-lorang",
-                    "username" => "bart-lorang"
-                ),
-                array(
-                    "bio" => "CEO & Co-Founder of @FullContact, Managing Director @v1vc_. Tech Entrepreneur, Investor. Husband to @parkerbenson and Father to Greyson Lorang",
-                    "followers" => 5454,
-                    "following" => 741,
-                    "type" => "twitter",
-                    "typeId" => "twitter",
-                    "typeName" => "Twitter",
-                    "url" => "https:\/\/twitter.com\/bartlorang",
-                    "username" => "bartlorang",
-                    "id" => "5998422"
-                ),
-            ),
-
-            "DigitalFootprint" => array(
-                "Topics" => array(
-                    array(
-                        "provider" => "aboutme",
-                        "value" => "Angel Investor"
-                    ),
-                    array(
-                        "provider" => "aboutme",
-                        "value" => "Entrepreneur"
-                    ),
-                    array(
-                        "provider" => "aboutme",
-                        "value" => "Husband"
-                    ),
-                    array(
-                        "provider" => "aboutme",
-                        "value" => "Tech Nerd"
-                    ),
-                    array(
-                        "provider" => "aboutme",
-                        "value" => "Technology"
-                    )
-                ),
-                "Scores" => ""
-            ),
-            "Organizations" => array(
-                array(
-                    "isPrimary" => false,
-                    "name" => "V1.vc",
-                    "startDate" => "2015-07",
-                    "title" => "Co-Founder & Managing Director",
-                    "current" => true
-                ),
-                array(
-                    "isPrimary" => false,
-                    "name" => "FullContact",
-                    "startDate" => "2010",
-                    "title" => "Co-Founder & CEO",
-                    "current" => true
-                )
-            ),
-            "Photos" => array(
-                array(
-                    "URL" => "https:\/\/upload.wikimedia.org/wikipedia\/en\/a\/a9\/Example.jpg",
-                    "Name" => "Facebook",
+        return [
+            "EmailAddresses" => [
+                "not-real@fake.com"
+            ],
+            "SocialMedia" => [
+                "Facebook" => [
+                    "url" => "facebook.com/not-real",
+                    "platform" => "Facebook",
+                    "handle" => "not-real"
+                ],
+                "LinkedIn" => [
+                    "url" => "linkedin.com/in/not-real",
+                    "platform" => "LinkedIn",
+                    "handle" => "not-real"
+                ],
+                "Twitter" => [
+                    "url" => "twitter.com/not-real",
+                    "platform" => "Twitter",
+                    "handle" => "not-real"
+                ]
+            ],
+            "PhoneNumbers" => [
+                "+44123456789"
+            ],
+            "Location" => [
+                [
+                    "Address" => [
+                        "PostalCode" => null,
+                        "Line1" => null,
+                        "Line2" => null,
+                        "Country" => "united kingdom",
+                        "Region" => "manchester"
+                    ]
+                ]
+            ],
+            "Name" => [
+                "FirstName" => "not",
+                "LastName" => "real",
+                "MiddleName" => null,
+                "FullName" => "not real"
+            ],
+            "Gender" => "male",
+            "Employment" => [
+                [
+                    "CompanyName" => "the fake company",
+                    "StartDate" => "2007-06",
+                    "EndDate" => "2007-10",
+                    "Title" => "venture capitalist",
                     "IsPrimary" => false
-                ),
-                array(
-                    "URL" => "https:\/\/test-of-tests-for-testing.s3-eu-west-1.amazonaws.com\/bart%40fullcontact.com\/image\/Foursquare.png",
-                    "Name" => "Foursquare",
+                ],
+                [
+                    "CompanyName" => "the sequel to the fake company",
+                    "StartDate" => "2010-02",
+                    "EndDate" => null,
+                    "Title" => "ceo",
                     "IsPrimary" => true
-                )
-            ),
-            "Bio" => "Mr. Lorang is a proven entrepreneur, executive and manager in the global technology industry. Mr. Lorang is active in the startup technology community as an angel investor, strategic advisor and speaker at industry events. Bart serves as Co-Founder & CEO of FullContact. Mr. Lorang is responsible for communicating FullContact's vision and strategy. Mr. Lorang is a visionary technologist with extensive experience conceiving, designing, building, marketing and selling enterprise software solutions on a global scale. Bart is also Co-Founder and Managing Director of v1.vc, a $5M seed stage fund dedicated to helping crazy entrepreneurs change the world. Bart serves on the Board of the Colorado Technology Association, Rapt Media and is on the Advisory Board of Education Funding Partners. Bart is a regular guest on FOX Business channel and has been featured by ABC, CNN, FOX News, MSNBC, Forbes, FastCompany, Yahoo, Inc Magazine and TechCrunch. Prior to founding FullContact, Mr. Lorang was an owner in Dimension Technology Solutions where he served as President and oversaw all day to day operations, customer engagements, partner relations, product development, sales and marketing functions. Mr. Lorang is recognized for providing solutions that are simple and work reliably. He strongly believes in using technology to solve problems as opposed to using problems to demonstrate technology. Mr. Lorang holds a Bachelor of Science degree in Computer Science from the University of Colorado and an MBA from the Daniels College of Business at University of Denver. Specialties: Investments, Startups, Financial Analysis, Sales, Technical Sales, Implementations, System Integration, Project Management, Leadership",
-            "JobTitle" => "CEO"
-        );
+                ]
+            ],
+            "Education" => [
+                [
+                    "Name" => "the fake school",
+                    "StartDate" => null,
+                    "EndDate" => "1996"
+                ]
+            ]
+        ];
     }
-
     public function testGetFirstName()
     {
         $testData = $this->_getTestData();
         $response = new Person($testData);
 
-        $strExpectedFirstName = "Bart";
+        $strExpectedFirstName = "not";
         $strReturnedFirstName = $response->getFirstName();
 
         $this->assertEquals($strExpectedFirstName, $strReturnedFirstName);
 
         // Test to make sure it returns null when there's no firstname
-        unset($testData["ContactInfo"]["FirstName"]);
+        unset($testData["Name"]["FirstName"]);
         $response = new Person($testData);
 
         $this->assertEquals(null, $response->getFirstName());
 
         // Test for if the ContactInfo field doesn't exist
-        unset($testData["ContactInfo"]);
+        unset($testData["Name"]);
+        $response = new Person($testData);
+
+        $this->assertEquals(null, $response->getFirstName());
+    }
+    public function testGetLastName()
+    {
+        $testData = $this->_getTestData();
+        $response = new Person($testData);
+
+        $strExpectedLastName = "real";
+        $strReturnedLastName = $response->getLastName();
+
+        $this->assertEquals($strExpectedLastName, $strReturnedLastName);
+
+        // Test to make sure it returns null when there's no LastName
+        unset($testData["Name"]["LastName"]);
+        $response = new Person($testData);
+
+        $this->assertEquals(null, $response->getLastName());
+
+        // Test for if the ContactInfo field doesn't exist
+        unset($testData["Name"]);
         $response = new Person($testData);
 
         $this->assertEquals(null, $response->getFirstName());
     }
 
-    public function testGetLastName()
-    {
-        $testData = $this->_getTestData();
-        $response = new Person($this->_getTestData());
-
-        $strExpectedLastName = "Lorang";
-        $strReturnedLastName = $response->getLastName();
-
-        $this->assertEquals($strExpectedLastName, $strReturnedLastName);
-
-        unset($testData["ContactInfo"]["LastName"]);
-        $response = new Person($testData);
-
-        $this->assertEquals(null, $response->getLastName());
-
-        unset($testData["ContactInfo"]);
-        $response = new Person($testData);
-
-        $this->assertEquals(null, $response->getLastName());
-    }
-    public function testGetBio()
-    {
-        $testData = $this->_getTestData();
-        $response = new Person($this->_getTestData());
-
-        $this->assertEquals($testData['Bio'], $response->getBio());
-
-        unset($testData['Bio']);
-        $response = new Person($testData);
-
-        $this->assertEquals(null, $response->getBio());
-    }
     public function testGetRole()
     {
         $testData = $this->_getTestData();
-        $response = new Person($this->_getTestData());
+        $response = new Person($testData);
+        $itemRole = $response->getRole();
 
-        $this->assertEquals($testData['JobTitle'], $response->getRole());
+        $this->assertTrue($itemRole->bPrimary());
+        $this->assertEquals('ceo', $itemRole->getTitle());
+        $this->assertEquals('the sequel to the fake company', $itemRole->getName());
+        $this->assertEquals('2010-02', $itemRole->getStartDate());
+        $this->assertEquals('', $itemRole->getEndDate());
 
-        unset($testData['JobTitle']);
+        unset($testData['Employment']);
         $response = new Person($testData);
 
         $this->assertEquals(null, $response->getRole());
@@ -263,7 +145,6 @@ class PersonTest
 
         $this->assertTrue(is_array($arrProfiles));
     }
-
     public function testGetPhotos()
     {
         $response = new Person($this->_getTestData());
@@ -272,50 +153,17 @@ class PersonTest
 
         $this->assertTrue(is_array($arrPhotos));
     }
-    public function testGetPrimaryPhoto()
+    public function testGetEducation()
     {
-        $response = new Person($this->_getTestData());
+        $testData = $this->_getTestData();
+        $response = new Person($testData);
 
-        $itemPhoto = $response->getPrimaryPhoto();
+        $arrEducation = $response->getEducation();
 
-        $this->assertTrue($itemPhoto->bPrimary());
-        $this->assertEquals("Foursquare", $itemPhoto->getName());
-        $this->assertEquals(
-            "https:\/\/test-of-tests-for-testing.s3-eu-west-1.amazonaws.com\/bart%40fullcontact.com\/image\/Foursquare.png",
-            $itemPhoto->getURL()
-        );
-    }
-    public function testGetPrimaryPhoto_NoPrimary()
-    {
-        $response = new Person([
-            "Photos" => [
-                [
-                    "URL" => "a URL",
-                    "Name" => "a Name",
-                    "IsPrimary" => false
-                ],
-                [
-                    "URL" => "another URL",
-                    "Name" => "another Name",
-                    "IsPrimary" => false
-                ]
-            ]
-        ]);
+        $this->assertTrue(is_array($arrEducation));
 
-        $itemPhoto = $response->getPrimaryPhoto();
+        $expectedResponse = $testData['Education'];
 
-        // Should pick first image
-        $this->assertFalse($itemPhoto->bPrimary());
-        $this->assertEquals("a Name", $itemPhoto->getName());
-        $this->assertEquals("a URL", $itemPhoto->getURL());
-    }
-    public function testGetPrimaryPhoto_NoPhotos()
-    {
-        $response = new Person([]);
-
-        $itemPhoto = $response->getPrimaryPhoto();
-
-        // Should return null
-        $this->assertNull($itemPhoto);
+        $this->assertEquals($expectedResponse, $arrEducation);
     }
 }
